Implement Gaussian Mixture Model.

+++ mleap-core/src/main/scala/ml/combust/mleap/core/clustering/KMeansModel.scala
@@ -17,9 +17,9 @@ object KMeansModel {
 case class KMeansModel(clusterCenters: Array[VectorWithNorm]) {
   def clusterCount: Int = clusterCenters.length
 
-  def apply(features: Vector): Double = predict(features)
+  def apply(features: Vector): Int = predict(features)
 
-  def predict(features: Vector): Double = {
+  def predict(features: Vector): Int = {
     findClosest(VectorWithNorm(features))._1
   }
